Breakout windows into separate modules. Auto-import current modules at pymeasure/display/windows/__init__.py
